[Dynamo] Code refactor: moving variables related unit tests to a separate folder
