+++ test/dynamo/test_repros.py
@@ -1862,17 +1862,17 @@ class ReproTests(torch._dynamo.test_case.TestCase):
 
     def test_relative_import(self):
         try:
-            from . import test_functions as _  # noqa: F401
+            from .variables import test_functions as _  # noqa: F401
 
             def fn(x):
-                from .test_functions import tensor_for_import_testing
+                from .variables.test_functions import tensor_for_import_testing
 
                 return x * 2 * tensor_for_import_testing
 
         except ImportError:
 
             def fn(x):
-                from test_functions import tensor_for_import_testing
+                from .variables.test_functions import tensor_for_import_testing
 
                 return x * 2 * tensor_for_import_testing
 
@@ -1885,17 +1885,17 @@ class ReproTests(torch._dynamo.test_case.TestCase):
 
     def test_relative_import_no_modulename(self):
         try:
-            from . import test_functions as _  # noqa: F401
+            from .variables import test_functions as _  # noqa: F401
 
             def fn(x):
-                from . import test_functions
+                from .variables import test_functions
 
                 return x * 2 * test_functions.tensor_for_import_testing
 
         except ImportError:
 
             def fn(x):
-                import test_functions
+                from .variables import test_functions
 
                 return x * 2 * test_functions.tensor_for_import_testing
