+++ test/dynamo/variables/test_global.py
