+++ test/dynamo/variables/test_global_declaration.py
